Conflicts:
	src/core/tomasulo/instruction.cpp

+++ include/tomasulo/reservation_station.hpp
@@ -24,7 +24,7 @@ private:
     unsigned int destination;
     uint16_t address;
     unsigned int delay_cycles;
-    unsigned int rob_id;
+    unsigned short rob_id;
 
 public:
     reservation_station(FUNCTION_UNIT fu,
@@ -35,7 +35,7 @@ public:
     void flush();
     STATE tick(tomasulo & t);
     STATE get_state(){return state;}
-    bool issue(Instruction inst, unsigned int rob_id);
+    bool issue(Instruction inst, unsigned short rob_id);
 };
 
 #endif
